Drawing lines should work properly now.
Certain variables were made static for the sake of getting things to
work.
Connecting to same color and in-range gems should now work.
The board doesn't reset, but the gem chain does, so you can continuously
draw lines.

Currently, RGBY gems classes are identical except in the Color, which is
painted by the engine.  Also the color detection is different since it
checks for class type instead of sprite.